move card styles to own file
